requirements.txt
aniso8601==7.0.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
firebase-admin==4.5.1
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.35.0
h11==0.12.0
httplib2==0.18.1
idna==2.10
msgpack==1.0.2
promise==2.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
python-dotenv==0.15.0
python-gql==0.2.3
pytz==2021.1
requests==2.25.1
rsa==4.7
Rx==1.6.1
six==1.15.0
starlette==0.13.6
starlette-graphql==0.1.5
uritemplate==3.0.1
urllib3==1.26.3
uvicorn==0.13.3
